Remove votesRouter and update cyclesRouter
