+++ src/routers/users.ts
@@ -1,7 +1,7 @@
 import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { default as express } from 'express';
 import type * as db from '../db';
-import { getUser, getUserAttributes, getVotes, updateUser } from '../services/users';
+import { getUser, getUserAttributes, updateUser } from '../services/users';
 import { getGroupsPerUser } from '../services/groups';
 import { isLoggedIn } from '../middleware/isLoggedIn';
 const router = express.Router();
@@ -9,7 +9,6 @@ const router = express.Router();
 export function usersRouter({ dbPool }: { dbPool: PostgresJsDatabase<typeof db> }) {
   router.get('/', isLoggedIn(), getUser(dbPool));
   router.put('/:userId', isLoggedIn(), updateUser(dbPool));
-  router.get('/:userId/cycles/:cycleId/votes', isLoggedIn(), getVotes(dbPool));
   router.get('/:userId/groups', isLoggedIn(), getGroupsPerUser(dbPool));
   router.get('/:userId/attributes', isLoggedIn(), getUserAttributes(dbPool));
   return router;
