+++ src/services/users.ts
@@ -1,7 +1,7 @@
 import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import * as db from '../db';
 import type { Request, Response } from 'express';
-import { and, desc, eq } from 'drizzle-orm';
+import { eq } from 'drizzle-orm';
 import { insertUserSchema } from '../types/users';
 import { overwriteUsersToGroups } from './usersToGroups';
 import { upsertUserAttributes } from './userAttributes';
@@ -54,37 +54,6 @@ export function getUserAttributes(dbPool: PostgresJsDatabase<typeof db>) {
   };
 }
 
-export function getVotes(dbPool: PostgresJsDatabase<typeof db>) {
-  return async function (req: Request, res: Response) {
-    const sessionUserId = req.session.userId;
-    const cycleId = req.params.cycleId;
-    const userId = req.params.userId;
-    if (userId !== sessionUserId) {
-      return res.status(400).json({
-        errors: [
-          {
-            message: 'Not authorized to query this user',
-          },
-        ],
-      });
-    }
-
-    if (!cycleId) {
-      return res.status(400).json({
-        errors: [
-          {
-            message: 'Expected cycleId in query params',
-          },
-        ],
-      });
-    }
-
-    const votesRow = await getVotesForCycleByUser(dbPool, userId, cycleId);
-
-    return res.json({ data: votesRow });
-  };
-}
-
 export function updateUser(dbPool: PostgresJsDatabase<typeof db>) {
   return async function (req: Request, res: Response) {
     // parse input
@@ -130,36 +99,4 @@ export function updateUser(dbPool: PostgresJsDatabase<typeof db>) {
 
     return res.json({ data: { user, updatedGroups, updatedUserAttributes } });
   };
-}
-
-export async function getVotesForCycleByUser(
-  dbPool: PostgresJsDatabase<typeof db>,
-  userId: string,
-  cycleId: string,
-) {
-  const response = await dbPool.query.cycles.findMany({
-    with: {
-      forumQuestions: {
-        with: {
-          questionOptions: {
-            with: {
-              votes: {
-                where: and(eq(db.votes.userId, userId)),
-                limit: 1,
-                orderBy: [desc(db.votes.createdAt)],
-              },
-            },
-          },
-        },
-      },
-    },
-    where: eq(db.cycles.id, cycleId),
-  });
-
-  const out = response.flatMap((cycle) =>
-    cycle.forumQuestions.flatMap((question) =>
-      question.questionOptions.flatMap((option) => option.votes),
-    ),
-  );
-  return out;
 }
