+++ src/services/votes.spec.ts
@@ -6,7 +6,7 @@ import { runMigrations } from '../utils/db/runMigrations';
 import { insertVotesSchema } from '../types';
 import { cleanup, seed } from '../utils/db/seed';
 import { z } from 'zod';
-import { insertVote } from './votes';
+import { insertVote, getVotesForCycleByUser } from './votes';
 import { eq } from 'drizzle-orm';
 
 const DB_CONNECTION_URL = 'postgresql://postgres:secretpassword@localhost:5432';
@@ -15,7 +15,11 @@ describe('service: votes', () => {
   let dbPool: PostgresJsDatabase<typeof db>;
   let dbConnection: postgres.Sql<NonNullable<unknown>>;
   let testData: z.infer<typeof insertVotesSchema>;
-  let cycleId: string;
+  let cycle: db.Cycle | undefined;
+  let questionOption: db.QuestionOption | undefined;
+  let forumQuestion: db.ForumQuestion | undefined;
+  let user: db.User | undefined;
+  let otherUser: db.User | undefined;
   beforeAll(async () => {
     const initDb = createDbPool(DB_CONNECTION_URL, { max: 1 });
     await runMigrations(DB_CONNECTION_URL);
@@ -24,17 +28,21 @@ describe('service: votes', () => {
     // seed
     const { users, questionOptions, forumQuestions, cycles } = await seed(dbPool);
     // Insert registration fields for the user
+    questionOption = questionOptions[0];
+    forumQuestion = forumQuestions[0];
+    user = users[0];
+    otherUser = users[1];
+    cycle = cycles[0];
     testData = {
       numOfVotes: 1,
-      optionId: questionOptions[0]?.id ?? '',
-      questionId: forumQuestions[0]?.id ?? '',
-      userId: users[0]?.id ?? '',
+      optionId: questionOption?.id ?? '',
+      questionId: forumQuestion?.id ?? '',
+      userId: user?.id ?? '',
     };
-    cycleId = cycles[0]?.id ?? '';
   });
 
   it('should save vote', async () => {
-    await dbPool.update(db.cycles).set({ status: 'OPEN' }).where(eq(db.cycles.id, cycleId));
+    await dbPool.update(db.cycles).set({ status: 'OPEN' }).where(eq(db.cycles.id, cycle!.id));
     // Call the saveVote function
     const { data: response } = await insertVote(dbPool, testData);
     // Check if response is defined
@@ -51,7 +59,7 @@ describe('service: votes', () => {
 
   it('should not save vote if cycle is closed', async () => {
     // update cycle to closed state
-    await dbPool.update(db.cycles).set({ status: 'CLOSED' }).where(eq(db.cycles.id, cycleId));
+    await dbPool.update(db.cycles).set({ status: 'CLOSED' }).where(eq(db.cycles.id, cycle!.id));
     // Call the saveVote function
     const { data: response, errors } = await insertVote(dbPool, testData);
 
@@ -64,7 +72,7 @@ describe('service: votes', () => {
 
   it('should not save vote if cycle is upcoming', async () => {
     // update cycle to closed state
-    await dbPool.update(db.cycles).set({ status: 'UPCOMING' }).where(eq(db.cycles.id, cycleId));
+    await dbPool.update(db.cycles).set({ status: 'UPCOMING' }).where(eq(db.cycles.id, cycle!.id));
     // Call the saveVote function
     const { data: response, errors } = await insertVote(dbPool, testData);
 
@@ -75,6 +83,49 @@ describe('service: votes', () => {
     expect(errors).toBeDefined();
   });
 
+  it('should get votes latest votes related to user', async function () {
+    // create vote in db
+    await dbPool.insert(db.votes).values({
+      numOfVotes: 2,
+      optionId: questionOption!.id,
+      questionId: forumQuestion!.id,
+      userId: user!.id,
+    });
+    // create second interaction with option
+    await dbPool.insert(db.votes).values({
+      numOfVotes: 10,
+      optionId: questionOption!.id,
+      questionId: forumQuestion!.id,
+      userId: user!.id,
+    });
+
+    const votes = await getVotesForCycleByUser(dbPool, user!.id, cycle!.id);
+    // expect the latest votes
+    expect(votes[0]?.numOfVotes).toBe(10);
+  });
+
+  test('should not get votes for other user', async function () {
+    // create vote in db
+    await dbPool.insert(db.votes).values({
+      numOfVotes: 2,
+      optionId: questionOption!.id,
+      questionId: forumQuestion!.id,
+      userId: otherUser!.id,
+    });
+    // create second interaction with option
+    await dbPool.insert(db.votes).values({
+      numOfVotes: 10,
+      optionId: questionOption!.id,
+      questionId: forumQuestion!.id,
+      userId: otherUser!.id,
+    });
+
+    // user 1 gets votes but it should not include otherUser votes
+    const votes = await getVotesForCycleByUser(dbPool, user!.id, cycle!.id);
+
+    // no votes have otherUser's id in array
+    expect(votes.filter((vote) => vote.userId === otherUser?.id).length).toBe(0);
+  });
   afterAll(async () => {
     // Delete registration data
     await cleanup(dbPool);
